% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolightviz.R
\name{geolightviz}
\alias{geolightviz}
\title{Start the GeoLightViz shiny app}
\usage{
geolightviz(x, path = NULL, launch_browser = TRUE, run_bg = TRUE)
}
\arguments{
\item{x}{a GeoPressureR \code{tag} object, a \code{.Rdata} file or the
unique identifier \code{id} with a \code{.Rdata} file located in \code{"./data/interim/{id}.RData"}.}

\item{path}{a GeoPressureR \code{path} or \code{pressurepath} data.frame.}

\item{launch_browser}{If true (by default), the app runs in your browser, otherwise it runs on
Rstudio.}

\item{run_bg}{If true (by default), the app runs in a background R process using \code{callr::r_bg()},
allowing you to continue using the R console. If false, the app blocks the console until closed.}
}
\description{
Start the GeoLightViz shiny app
}
